FlyTest_Django/requirements.txt
# Core framework
Django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
django-cors-headers==4.3.1
python-dotenv==1.1.1
django-filter==25.1
drf-nested-routers==0.94.2

# Async and WebSocket
channels==4.3.2
channels-redis==4.3.0
daphne==4.2.1

# OpenAI and agent stack
openai>=1.109.1
langgraph==1.0.6
langchain==1.2.3
langchain-openai==1.1.7
langchain-qwq==0.3.4
langchain-core==1.2.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk>=0.2.0
langsmith>=0.3.45,<1.0.0
langgraph-checkpoint-sqlite==3.0.2
langgraph-checkpoint-postgres==3.0.3
langchain-mcp-adapters==0.2.1

# Optional providers kept for reference
# langchain-anthropic>=0.3.0
# anthropic>=0.48.0
# langchain-experimental>=0.4.0
# langchain-huggingface==0.2.0
# langchain-cohere==0.3.5
# cohere==5.15.0
# langchain-fireworks==0.2.7
# fireworks-ai==0.15.12
# langchain-google-genai==2.0.8
# google-ai-generativelanguage==0.6.15
# google-api-core==2.24.1
# google-api-python-client==2.162.0
# google-auth==2.38.0
# google-generativeai==0.8.4
# langchain-mistralai==0.2.4

# Database drivers
psycopg2-binary==2.9.10
psycopg[binary]==3.3.1

# Spreadsheet and office documents
openpyxl==3.1.5
pandas==2.2.3
xlrd==2.0.1
pypdf==6.9.2
python-docx==1.2.0
python-pptx==1.0.2
docx2txt==0.9
olefile>=0.46
markdown==3.9
unstructured==0.17.2
beautifulsoup4==4.13.3

# Knowledge base and vector search
langchain-community==0.4.1
langchain-text-splitters==1.1.0
langchain-qdrant==1.0.0
qdrant-client==1.11.3
fastembed>=0.7.0,<1.0
langdetect==1.0.9

# Optional local embedding stack
# sentence-transformers==4.1.0
# torch==2.7.1
# transformers==4.52.4
# huggingface-hub==0.32.4

# Background jobs and service runtime
celery==5.4.0
redis==5.2.0
uvicorn[standard]==0.34.0
httpx==0.28.1

# Test automation and utilities
pytest>=8.0.0
playwright>=1.40.0
browser-use==0.10.1
Pillow>=10.0.0
Faker>=28.4.1
jsonpath-ng>=1.7.0
croniter>=3.0.3
qrcode>=8.2
python-barcode>=0.15.1
pycryptodome>=3.21.0
PyYAML>=6.0.2
tiktoken>=0.5.0
nltk>=3.8.0
websocket-client>=1.6.0
